pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "researcher_agent"
version = "0.1.0"
description = "A parallelized, AI-driven research assistant."
readme = "README.md"
authors = [
  { name = "Mark Pesce", email = "mark@markpesce.com" },
]
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.10"
dependencies = [
    "google-generativeai",
    "playwright",
    "googlesearch-python",
    "pandas",
    "pydantic",
    "python-dotenv",
    "colorama",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "mypy",
    "types-requests",
    "black",
    "isort"
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = [".", "src"]
